External Audit Storage: add usage event (#34915)
* External Audit Storage: add usage event

This adds a regular usage event every time credentials are refreshed for
the External Audit Storage feature.

It's desirable to regularly submit an event instead of just a one-off
when the feature is turned on, because there's no guarantee an
individual event will actually make it to prehog/posthog and won't be
lost. This event should be emitted roughly every 45 minutes.

* send empty event

Co-authored-by: Edoardo Spadolini <edoardo.spadolini@goteleport.com>

---------

Co-authored-by: Edoardo Spadolini <edoardo.spadolini@goteleport.com>