[v14] drone: Reorder pipelines to fix dependency
Move the recently added spacelift promotion pipeline before the
`publish-rlz` pipeline as it has a dependency on it, and maybe drone
needs these things ordered like this. We are currently seeing the error

    linter: invalid or unknown pipeline dependency

on the drone web ui since the spacelift changes to `.drone.yml` were
merged. Let's see if this fixes that.