Fix incorrect Kubernetes event watcher chunking
Improve handling of streaming events in response to non-chunked content-type from the Kubernetes API.

Previously, when dealing with streaming events, the code streamed the response into the target connection.
However, the Kubernetes API did not provide the content-type as chunked.
This lack of information caused the forwarder to be unaware that the connection was chunked.
Consequently, the forwarder delayed response writing by buffering it, intending to minimize the number of writes.

In scenarios where the connection stream was busy with events, users might not receive individual events as chunks,
leading to incomplete data. This could result in users receiving malformed JSON, triggering an abort of the process.

To address this issue, this commit introduces a modification. After each event, the response is now flushed to ensure
that users receive the event as a complete and well-formed chunk.
This adjustment improves the reliability and integrity of the data received by users during streaming events.

Fixes #34977

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>