Metrics: add running services metric
This PR adds a new gauge metric: `teleport_services`
This metric has a label identifying the service and whether or not it is
running.
Those services are the ones started in the supervisor.
Eg, proxy.web, discovery.init, ssh.node, auth.tls

When the service stops, the counter is decreased.

This gives us an overview of the currently running services in the
process.