Improve the agent-pool-terraform module
Responds to feedback by **r0mant**.

Add an input variable for the user's Teleport edition and use that input
variable to supply the second parameter of the one-line install script,
which specifies the Teleport edition. This way, we do not need to branch
the user data script within the module.

This change also includes the following edits, which were required to
test the guide:

- Put the version number in quotes within `main.auto.tfvars`.
- Add the `terraform init` command.
- Add an AMI data source instead of hardcoding an AMI ID. This prevents
  the AMI ID from going out of date.

For neatness, this also adds a separate `.tf` file for the token,
preventing the file from appearing too long when we include it in the
guide.