build: Deprecate centos7 release targets
Deprecate the centos7 release targets in `build.assets/Makefile`, by
making them just depend on the main release targets with no other
specific functionality.

That some builds use centos7 is an implementation detail that does not
need to be exposed via the "public" targets of the Makefile. We just
want to be able to build amd64 and amd64-fips releases, regardless of
what is used to build that. We do not build or release non-centos7
binaries so this distinction just causes confusion.

This moves the buildbox dependency for the fips build to the generic
target such as it is already for the non-fips build.