build: Repair release-arm64 release target
Repair the `release-arm64` release target in `build.assets/Makefile` so
that it can be called from the common GitHub Actions workflow for
building linux releases. This performs the same build as the specific
build-linux-arm64 workflow which currently calls the internal build
targets directly.

Add a check to the release and buildbox targets to ensure the host
architecture matches the target build architecture as the arm64 build
cannot be cross-compiled currently. CI ensures that the correct runner
architecture is used for the build.

Update `build-package.sh` to use a different version of `fpm` for
building deb/rpm archives as the new version has an arm64 variant that
is needed for the GitHub Actions workflow to build when running on
arm64.