build: Remove BUILDBOX_NAME variable from build.assets
Remove the `BUILDBOX_NAME` variable from `build.assets/Makefile` and
instead parameterize the buildbox target for the `release` target.

`BUILDBOX_NAME` was used to short-circuit the `buildbox` target so as to
not build if we were building a buildbox different to `$(BUILDBOX)`. For
the dependencies to work for `release-arm`, the `release-arm` target had
an explicit dependency on `buildbox-arm`.

This is cleaned up by having the `release` target depend on a dynamic
buildbox target based on `$(BUILDBOX)`. There is only one special case
for this, `$(BUILDBOX_ARM)`, but more can be added in future very
easily by defining the mapping between the specific buildbox and the
target used to make it.