build: Pull buildboxes on all CI, not just Drone
Before building a buildbox, pull the buildbox image to allow the build
to go quicker if little or nothing has changed, taking advantage of
caching. This was already done for Drone, but is now expanded for any CI
run if `$CI` is set.

Use a make "macro" to make it a little clearer what is going on.