Use github converter in gRPC APIs
Any github auth connectors created from an enterprise build would
be created but failed to be returned in Create/Update because the
type assertions don't know about the enterprise version of the
resource. All instances in the gRPC client and server now use the
`services.ConvertGithubConnector` mechanism instead of direct type
assertions.