Continue completing uploads on not found errors (#35102)
Prior to this change, if we failed to complete an upload, the completed
would error out and stop processing uploads until the next interval
(5 minutes later). This leads to us falling behind and allowing a large
number of uncompleted uploads to accumulate, particularly when there
are multiple auth servers attempting to complete the same set of
uploads.

Fixes gravitational/cloud#5938