drone: Fix signature (#35109)
The current drone.yml is causing drone to ask for approval for builds.
This is because the drone signature is incorrect.

eb7362962368cb1b9156501aef2667e6c1ee0e14 updated it on the PR correctly
at first, but followed up with another that got the wrong signature.

The signature was generated with:

    make dronegen

and

    drone sign --save gravitational/teleport

Both generated the same sig/hash, in this commit.