kube-agent-updater: Add support for keyless signing
Keyless signing is being added as an additional signature method
in addition to the existing static signing key. Static key
signatures will be deprecated at some future point, but we'll be
supporting both for a while.

Also, we now publish signatures to the tlog due to upgrading to
cosign 2.x, so stop ignoring it.