ci: Support arm64 linux builds in common workflow (#2754)
Select a self-hosted arm64 runner when building arm64 linux releases as
we cannot cross-build arm64 binaries from amd64, which is the host
architecture of GitHub linux runners. This is what `release-linux-arm64`
currently does, and this change is to move the build from that workflow
to here.

Add filename patterns for arm64 to register with the release server so
that we provide the correct architecture.

Source-commit: gravitational/teleport.e@6f7460714581d87e6c5843e742837c1cbe74a4e1
Source-pull-requests: gravitational/teleport.e#2754