build: Update build.assets/Makefile to build arm64 (#35026)
* build: Clean up build.assets/Makefile

Do a little bit of cleaning on `build.assets/Makefile`:
* Remove unnecessary semicolons on some recipes
* Fix quoting on a few targets, some where double quotes were missing,
  one where there was a stray double quote
* Reorder the release targets to put the ones that should be used from
  outside up front and the generic internal targets after that
  (essentially moving the `release` target to go with the other generic
  release targets) and add some comments explaining some of this.

This clean up will make subsequent commits a little easier to review due
to less code movement in those commits.

* build: Deprecate centos7 release targets

Deprecate the centos7 release targets in `build.assets/Makefile`, by
making them just depend on the main release targets with no other
specific functionality.

That some builds use centos7 is an implementation detail that does not
need to be exposed via the "public" targets of the Makefile. We just
want to be able to build amd64 and amd64-fips releases, regardless of
what is used to build that. We do not build or release non-centos7
binaries so this distinction just causes confusion.

This moves the buildbox dependency for the fips build to the generic
target such as it is already for the non-fips build.

* build: Repair release-arm64 release target

Repair the `release-arm64` release target in `build.assets/Makefile` so
that it can be called from the common GitHub Actions workflow for
building linux releases. This performs the same build as the specific
build-linux-arm64 workflow which currently calls the internal build
targets directly.

Add a check to the release and buildbox targets to ensure the host
architecture matches the target build architecture as the arm64 build
cannot be cross-compiled currently. CI ensures that the correct runner
architecture is used for the build.

Update `build-package.sh` to use a different version of `fpm` for
building deb/rpm archives as the new version has an arm64 variant that
is needed for the GitHub Actions workflow to build when running on
arm64.

* build: Remove BUILDBOX_NAME variable from build.assets

Remove the `BUILDBOX_NAME` variable from `build.assets/Makefile` and
instead parameterize the buildbox target for the `release` target.

`BUILDBOX_NAME` was used to short-circuit the `buildbox` target so as to
not build if we were building a buildbox different to `$(BUILDBOX)`. For
the dependencies to work for `release-arm`, the `release-arm` target had
an explicit dependency on `buildbox-arm`.

This is cleaned up by having the `release` target depend on a dynamic
buildbox target based on `$(BUILDBOX)`. There is only one special case
for this, `$(BUILDBOX_ARM)`, but more can be added in future very
easily by defining the mapping between the specific buildbox and the
target used to make it.

* build: Pull buildboxes on all CI, not just Drone

Before building a buildbox, pull the buildbox image to allow the build
to go quicker if little or nothing has changed, taking advantage of
caching. This was already done for Drone, but is now expanded for any CI
run if `$CI` is set.

Use a make "macro" to make it a little clearer what is going on.