[v13] build: Update build.assets/Makefile to build arm64
Update `build.assets/Makefile` so that the `release-arm64` target works
again so that it can be called from the release-linux GitHub Actions
workflow.

There are multiple steps to this, cleaning up some existing cruft along
the way as there were some easy wins once the refactoring for the arm64
build were done.

There were also a few targets with some mismatched quotes in the
recipies that have been fixed, and some old argument no longer necessary
in the `buildbox-arm` target.

This change will facilitate moving the deb/rpm parts of the arm64 build
from Drone to GitHub Actions - currently the tarballs are built on GHA
using a workflow just for ARM64. The generic linux build workflow builds
the deb/rpm packages but requires a target in `build.assets/Makefile` to
perform the release build.