Assign reviewer on post-release PRs (#35158)
Now that our post-release PRs are automated, the individual publishing
the release only has to look out for them to be posted and help push
them through. This often gets forgotten, and the post release work
doesn't get merged in a timely manner. Fix this by ensuring that the
individual who published the release is automatically tagged as a
reviewer of the PRs.