auth: fix panic with v14 clients and sync recording modes (#35183)
In #34793 we added a fix for missing audit events when v13 agents
use sync recording modes against a v14 cluster. This fix missed a
few events that used to be filtered out by TeeStreamer, and as a
result we ended up writing session recording data for desktop and
app sesions to the audit log. This data eventually caused a panic,
likely due to an integer overflow.