feat: Create endpoint for EAS and region checks (#2776)
This commit adds a CreateDraftExternalAuditStorage endpoint to be used
by tctl in https://github.com/gravitational/teleport/pull/35166 to avoid
overwriting the current draft without -f

It also adds a check that the region of the new draft matches the region
of the current cluster audit config. This is currently a requirement
for External Audit Storage that the infra be stored in the same region
at the Auth server, which we get via the audit config.

Source-commit: gravitational/teleport.e@7c609782133d1caf8d55686af2666b8e37153906
Source-pull-requests: gravitational/teleport.e#2776