Include Machine ID instructions in plugin guides
Closes #15062

This change incorporates Machine ID instructions into how-to guides for
Teleport plugins.

The alternative was to write a separate guide for using Machine ID with
plugins. However, the current how-to guides assume that the user is
providing long-lived credentials via `tctl`, so it makes sense to modify
the instructions in these guides. Otherwise, a user would need to get
familiar with the Machine ID guide, then work out for themselves which
parts of the how-to guides are relevant or not.

- Use a partial for the initial plugin config as well as for a Machine
  ID prerequisite.
- Add partial for providing identity files via Machine ID.
- Edit impersonation partials to mention Machine ID.
- Add a `refresh-plugin-identity.mdx` partial for configuring plugins to
  periodically refresh their identities.
- Make Access Request plugin guides more consistent with how they use
  the identity-export.mdx partial.