fix: External Audit Storage bootstrap in us-east-1 (#35324)
This commit fixes the S3 bucket creation for External Audit Storage when
the region is us-east-1.
It's currently broken because we unconditionally set a location
constraint on the bucket so that it will be in our desired region.
This fails when the region is us-east-1.
Because that's the default region for S3 buckets, and AWS has decided
that explicitly asking for the default is wrong.

Changelog: fixes resource bootstrap for External Audit Storage in us-east-1.