Default desktop port to 3389 if not specified (#35345)
When specifying hosts in the config file, Teleport
will automatically set the port to 3389 if it is
not otherwise specified. This behavior is different
with tctl or our API - in these cases, we don't
default to the correct port, and attempts to connect
to these desktops will fail.

We now parse the desktop addr at connect time and
set the port if it is not provided.

Closes #31646