[v14] Fix the IsAssistEnabled response (#35363)
* Fix the IsAssistEnabled response

A condition was added in the IsAssistEnabled function to check the license status of the Assist feature. Now, the function explicitly checks if the Assist feature is enabled in the license. This makes the check to be more explicit as the current implementation returns true only base on the API key and ACL control the feature access.

* Fix test