Attempt to fix `TestCreateResources` flakiness
The auth process created in the test was not being closed before
the test terminated leading to potentially writing data to the
temporary directory generated from `t.TempDir` while the testing
framework was trying to remove the directory. `TestEditResources`
also exhibited the same behvarior and was fixed here as well.