drone: Remove push-build-linux-arm64 pipeline (#2844)
Remove the push-build-linux-arm64 pipeline as it cannot work on drone.
It requires an ARM64 host to run and we do not have those drone runners.

This changed when the teleport/build.assets/Makefile was updated to run
the linux arm64 build the same way as was being done in the
build-linux-arm64 workflow. The old way in the makefile was not used in
any release so there's not much point running it in a push build
pipeline.

Source-commit: gravitational/teleport.e@c46ceffa5e8d7382cfd029bad01deb90e404fe4f
Source-pull-requests: gravitational/teleport.e#2844