Refine explanation and structure of access-graph.mdx
The Teleport Access Graph documentation is updated to improve clarity and accuracy. It now explains more clearly how it visualizes access to infrastructure, expands the main components from five to eight, and refines the explanation on User Groups. Additionally, the Access Graph section is relocated within the documentation structure for better organization.

Signed-off-by: Jakub Nyckowski <nyckowski.jakub@gmail.com>