Fix panic on nil value in `getPresetRoles()` (#35458)
* Fix panic on nil value in `getPresetRoles()`

This fixes a panic when attempting to fetch `/webapi/presetroles` as
`ui.NewRoles()` is (reasonably) not nil safe. `GetPresetRoles()` can
return nil entries if corresponding features are disabled, so this
tweaks `GetPresetRoles()` to filter out nil entries.

This endpoint is only used for Cloud features (email invites at
onboarding, specifically) and currently all features are enabled for
Cloud users, so the only internal use is minimally impacted.

The other use of `GetPresetRoles()`, `createPresetRoles()`, already
checks for nil entries. It's behavior is unchanged.

* Replace manual deletion with `slices.DeleteFunc`

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>

---------

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>