Prevent panic in web app transport (#35525)
The case where no app servers were available was not handled and
resulted in exceeding the bounds of the app servers slice. When
no servers are available now the slice is set to nil instead of
trying to shift it beyond the bounds. A test was added to validate
that the panic is not possible any more.

Fixes #35460