[v13] build: Fix dockerized arm build (#35467)
Fix the dockerized arm build by not overriding `BUILDBOX` as the build
of the arm buildbox (`buildbox-arm` target) requires that to not be
modified. Instead add a new `RELEASE_BUILDBOX` variable to specify which
buildbox used by the `release` target.

This fixes the push-build-linux-arm pipeline on v13 and v12. This is not
needed on v14 as it does the arm buildbox differently and does not
depend on the `BUILDBOX` variable being unchanged.