build: Revert Debian 12 bump for arm (32-bit) buildbox
Revert part of #31620 that updated the arm 32-bit buildbox from Debian
11 to Debian 12, rolling back to Debian 11. This increased the required
version of glibc above what we say we support. The update of the
distroless image is fine as that is self-contained. The buildbox builds
the stand-alone binaries and OS packages so must be conservative in that
it uses.

Reverts: https://github.com/gravitational/teleport/pull/31620 (partial)
Issue: https://github.com/gravitational/teleport/issues/35476