Prevent flakiness in TestOutput (#35558)
Tests were failing if too much time had passed between the output
of each logger being generated. The InDelta assertion was supposed
to prevent this, but because it used nanoseconds the delta of 10
was always exceeded and resulted in errors like below.

```
Max difference between 1701993842000000000 and 1701993843000000000 allowed is 10, but difference was -1e+09
```

To avoid this the assertion is now using Unix instead of UnixNano
so that we allow up to 10 seconds of drift in the timestamps.