Safer logging through `String()` defined in the interface
This commit adds `String()` to the interface of several structs which may contain sensitive data.  This is to provide compiler enforcement (and docs) around the need to ensure that sensitive values are not accidentally logged.

There was no cases of actual sensitive logging found, but this change will help ensure that regressions don't accidentally develop while also addressing the code scanning alerts recently found:
* https://github.com/gravitational/teleport/security/code-scanning/862
* https://github.com/gravitational/teleport/security/code-scanning/869
* https://github.com/gravitational/teleport/security/code-scanning/870
* https://github.com/gravitational/teleport/security/code-scanning/896
* https://github.com/gravitational/teleport/security/code-scanning/897
* https://github.com/gravitational/teleport/security/code-scanning/898