[v14] Adjust diagnostic messages of connection test for Connect My Computer (#35649)
* Adjust diagnostic messages of connection test for Connect My Computer

* Use JSDocs in services/agents/types.ts

* TextIcon: Align icon to top rather than center

When showing multiple TextIcons one after another, like when showing
connection diagnostic traces, it makes it easier to see where one
trace starts and another ends. Especially if one of the traces spans
multiple lines.

* Preserve line breaks from trace messages in UI

* Add ResourceTile to TestConnectionRequest

* Adjust diagnostic messages for Connect

* Remove LegacyTestConnection

* Offer better diag msg depending on SSH principal selection mode

* Suggest systemctl/launchctl depending on platform

* Rename ResourceTile to SSHNodeSetupMethod

* Add tests for launchctl vs systemctl

* Fix test after removal of "Linux" from "valid Linux login"

* Add tests for SSHConnectionTester

* Document valid values for SSHNodeSetupMethod

* Remove test for principal with no OS permissions to run as

In CI, it looks like we can absolutely spawn a new session as root, so
the test is meaningless there.

Instead, I'm going to add authhandlers tests and this test to the Connect
My Computer test plan.

* Adjust connectivity failed error expectation to v14 messaging