ci: Fix checkout-teleport action for private repos (#2901)
Fix the `checkout-teleport` action to work again with private teleport
repositories (teleport-private) which broke in a28ea2c4f when the
`tibdex/github-app-token` app was changed to
`actions/create-github-app-token`. The `repository` parameter was
changed to `repositories` where the values should no longer have the
organisation prefixed - i.e. `teleport-private` and not
`gravitational/teleport-private`.

To fix this we use the initial step that was used to validate the inputs
to also split the input repository to get just the repo component and
emit that as an output. While we're here, simplify the logic for
validating required inputs and add more validation for the values.

Update the README to fix the description of where the GitHub App needs
to be installed.

Source-commit: gravitational/teleport.e@8c47d6bccf951624d3b4db8b559071961d357c90
Source-pull-requests: gravitational/teleport.e#2901