[v14] Support paid editions in agent-pool-terraform (#35565)
* Support paid editions in agent-pool-terraform

Closes #33652

Currently, the `agent-pool-terraform` module only supports installing
Teleport Community Edition on agent nodes, which is not compatible with
commercial editions of Teleport.

This change edits the `agent-pool-terraform` module to support
commercial Teleport editions by:

- Adding a variation of the user data script that installs Teleport
  Enterprise
- Branching on the user data script to load based on an input variable

* Linter fixes

* Improve the agent-pool-terraform module

Responds to feedback by **r0mant**.

Add an input variable for the user's Teleport edition and use that input
variable to supply the second parameter of the one-line install script,
which specifies the Teleport edition. This way, we do not need to branch
the user data script within the module.

This change also includes the following edits, which were required to
test the guide:

- Put the version number in quotes within `main.auto.tfvars`.
- Add the `terraform init` command.
- Add an AMI data source instead of hardcoding an AMI ID. This prevents
  the AMI ID from going out of date.

For neatness, this also adds a separate `.tf` file for the token,
preventing the file from appearing too long when we include it in the
guide.

* Respond to hugoShaka feedback

- Disable non-SSH services by default on the agent token and in the
  configuration file written by the init script. Use comments to
  instruct the reader to enable other services.
- Add instructions in the how-to guide to enable services you want to
  run on agent nodes.

Also removes a statement about the init script that is no longer
accurate.

* Respond to webvictim feedback

Fix spelling