ci: Fix names of secrets needed to checkout teleport (#2917)
Fix the names of the secrets used when calling the `checkout-teleport`
action. The lower-cased names that were there come from workflows where
the secrets were passed in as input parameters. Instead just reference
the secrets as set up in the repository/environment directly.

This fixes the ability to build teleport-private as that is what the
secrets are needed for. This was not noticed before as we were always
checking out the public teleport repository for which secrets are not
needed.

Source-commit: gravitational/teleport.e@b2b1911ac0b01b2db6b0726b41b7a77b46c7ad05
Source-pull-requests: gravitational/teleport.e#2917