ci: Sync time on macOS runners (#2921)
* ci: Sync time on macOS runners

Synchronize time on the macOS runners to time.windows.com as we are
seeing a lot of clock skew affecting things like signing and token
usage:

    'Issued at' claim ('iat') must be an Integer representing the time that the assertion was issued
    [@octokit/auth-app] GitHub API time and system time are different by -158 seconds. Retrying request with the difference accounted for.

* ci: Remove arm64 beta macos runner fixes

Remove the fixes that chown a couple of directories and that install
`pkg-config` from the build-mac workflows. These have been fixed in the
runners for a while now and the fixes have been released.

Link: https://github.com/actions/runner-images/pull/7719
Link: https://github.com/actions/runner-images/commit/1bf18b5a8a5b4b91d597f6b1109b9ccc07ea6f4b

Source-commit: gravitational/teleport.e@6a9b607107965fcb97c07a8aa4a3e572a2e26d55
Source-pull-requests: gravitational/teleport.e#2921