ci: Add `path` parameter to checkout-teleport action (#2919)
* ci: Add `path` parameter to checkout-teleport action

Add a `path` parameter to the `checkout-teleport` action to specify a
directory where the repository should be checked out. Some of our
workflows check out into a subdirectory so as to keep other build
resources out of the repository, and they cannot use this action without
being able to specify a path.

* ci: Remove input settings that are default

Remove settings from the input parameters for values that are the
default - `required: false`, `default: ""`. No need to specify default
values.

Source-commit: gravitational/teleport.e@88d33232cf68f113643a7833d1b4b7bc05d42880
Source-pull-requests: gravitational/teleport.e#2919