ci: Use checkout-teleport action in oci workflows (#2920)
Use the updated `checkout-teleport` action that supports the `path`
parameter in the `oci-image-build` and `release-teleport-oci-distroless`
workflows instead of having the steps "open-coded". This allows us to
take advantage of the fix applied to the action with the `repositories`
parameter to `actions/create-github-app-token` in 6f98b29f59 and not
have to perform the same fix in these workflows.

Source-commit: gravitational/teleport.e@fc78d180dacd954d5e9f3d2bdde1d0762431472c
Source-pull-requests: gravitational/teleport.e#2920