Include user groups in unified app resources (#35697)
Previously when fetching apps, we would have to pass in a map of
all fetched apps mapped to all fetched user groups and then parse
that per app when making the ui struct. Now that we pass in an
individual app with unified resources, we can just get the user group
directly from the app and then parse from the user group lookup easier.