ci: Log into public ECR for some OCI builds (#2965)
* ci: Add public ECR login step to oci-image-build

Add an optional step to the oci-image-build workflow to log into public
ECR to raise rate limits for pulling images from there. Without logging
in, the rate limit is 1/second which apparently is not enough for some
of the builds (we certainly do not average anywhere near that pull rate,
so there is likely no burst capacity on the rate limit and docker does
not retry a pull).

The optional step is activated by providing a role name that has
permissions to pull from public ECR. Leaving the role name empty for the
input parameter will cause the workflow not to log into public ECR as
not all image builds need this.

Typically a caller of the oci-image-build workflow would pass
`${{ vars.ECR_READ_ONLY_ROLE }}` as the ARN which is available in the
`build`, `promote`, and `build-stage` GitHub environments. It will need
to be added to the others as we switch to them. However any role with
similar permissions to
`arn:aws:iam::146628656107:role/tf-teleport-build-read-only-role-gha`
could be used.

Typically this step is needed to pull the Teleport buildbox images.

* ci: Log into public ECR for kube-agent oci build

Add a parameter when calling `oci-image-build` in the kube-agent-updater
build workflow to log into public ECR. This is needed as that build
pulls the buildbox from public ECR and that is subject to low rate
limits when not authenticated.

We need to add an environment input parameter so we can access an
environment with a role name that has permissions to pull from public
ECR. We could hard-code a role name or an environment name, but we will
want to have all workflows parameterized with the environment ultimately
so we can easily do staging builds, so we'll just do that now.

Note that we cannot use `environment:` with `uses:` in the build job
(thanks GitHub), so we need to emit the value from the GitHub
environment that we want (`ECR_READ_ONLY_ROLE`) in the previous
`configure` job as an output so we can pass that value to
`oci-image-build`.

Fix up a little bit of shell quoting while we're here so that
`actionlint` stops complaining.

* ci: Log into public ECR for spacelift oci build

Add a parameter when calling `oci-image-build` in the spacelift
build workflow to log into public ECR. This is needed as that build
pulls the buildbox from public ECR and that is subject to low rate
limits when not authenticated.

We need to add an environment input parameter so we can access an
environment with a role name that has permissions to pull from public
ECR. We could hard-code a role name or an environment name, but we will
want to have all workflows parameterized with the environment ultimately
so we can easily do staging builds, so we'll just do that now.

Note that we cannot use `environment:` with `uses:` in the build job
(thanks GitHub), so we need to emit the value from the GitHub
environment that we want (`ECR_READ_ONLY_ROLE`) in the previous
`configure` job as an output so we can pass that value to
`oci-image-build`.

Fix up a little bit of shell quoting while we're here so that
`actionlint` stops complaining.

Source-commit: gravitational/teleport.e@c56b8c3309bc6ebb7edba787c9b529bb05cf380b
Source-pull-requests: gravitational/teleport.e#2965