Sort CLI reference guides alphabetically
Closes #21164

The CLI reference is a section of the docs in which each CLI tool has
its own page, so this change sorts each page individually.

The change places H2 sections not related to individual commands (e.g.,
sections for global flags and environment variables) at the top of each
guide, and sorts the sections following these.

In `teleport.mdx`, all commands except for `start` are only documented
in a table at the beignning of the page, so this change sorts the table.