Fix terminal tests
The testing utilities used to mimic the web ui half of the websocket
made assumptions that the session metadata would be the first message
provided. However, now that hostname resolution occurs after the
initial SSH connection has been established it is possible for other
messages to be sent first. This caused some tests to fail and others
to hang forever.

The duplicated terminal logic from (WebSuite) makeTerminal and
(testProxy makeTerminal has now consolidated into a single test
terminal utility. The new `terminal` now wraps a TerminalStream
which allows test code to make use of the message processing logic
already in place there instead of having to write custom logic
per test. All a test needs to do is provide handlers for any messages
that it desires to introspect.