Prevent NotFound Okta apps and groups from failing Okta assignment cleanup. (#2975)
* Prevent NotFound Okta apps and groups from failing Okta assignment cleanup.

If an Okta app or group is not found and an assignment referencing those apps
or groups is being cleaned up, the assignment client will now proceed as if
the assignment were successful. Since the app or group no longer exists,
there's not a whole lot we can do with it here.

* Use errors.As, use structured logging.

* Log created beforehand.

Source-commit: gravitational/teleport.e@7226024a2cf9335b0de66c293aa2f608785b0b5d
Source-pull-requests: gravitational/teleport.e#2975