Fix remote cluster active connection helper
WaitForActiveTunnelConnections would return if a the site was not
fully discovered yet which would result in test failures with a
message similar to the following:

```
failed connecting to host 127.0.0.1:35391: failed to receive cluster details response
        	            		failed to dial target host
        	            		unable to fetch node watcher, this proxy clusterPeer(TunnelConnection(name=00000000-0000-0000-0000-000000000000-cluster-remote, type=, cluster=cluster-remote, proxy=00000000-0000-0000-0000-000000000000)) has not been discovered yet, try again later
```

This occurred because the check was only verifying that the cluster
existed and was online, however the clusterPeer placeholder reports
a status of online even though it isn't usable. To prevent this the
check now asserts that the cluster is both online and is usable by
checking that the client for the site doesn't return an error. This
will prevent any clusterPeer objects from being considered in the
check since they always return an error when trying to get a client.